🎨 Moved the secure endpoints to /secure
