+++ middleware/app/api/main.py
@@ -1,13 +1,9 @@
-from api.routes import (developer, download, feedback, health, scan, subscribe,
-                        upload, user)
+from api.routes import download, health, upload
+from api.routes.secured.main import secured_api_router
 from fastapi import APIRouter
 
 api_router = APIRouter()
-api_router.include_router(health.router, prefix="/health", tags=["health"])
-api_router.include_router(upload.router, prefix="/upload", tags=["upload"])
-api_router.include_router(download.router, prefix="/download", tags=["download"])
-api_router.include_router(feedback.router, prefix="/feedback", tags=["feedback"])
-api_router.include_router(subscribe.router, prefix="/subscribe", tags=["subscribe"])
-api_router.include_router(user.router, prefix="/user", tags=["user"])
-api_router.include_router(scan.router, prefix="/scan", tags=["scan"])
-api_router.include_router(developer.router, prefix="/developer", tags=["developer"])
+api_router.include_router(health.router, prefix="/health", tags=["Health Routes"])
+api_router.include_router(upload.router, prefix="/upload", tags=["Upload Routes"])
+api_router.include_router(download.router, prefix="/download", tags=["Download Routes"])
+api_router.include_router(secured_api_router)
