--- ui/src/app/(pages)/help/page.tsx
@@ -1,16 +1,0 @@
-import { Metadata } from 'next'
-
-export const metadata: Metadata = {
-  title: 'Help',
-  description: 'Help page',
-}
-
-function HelpPage() {
-  return (
-    <div className="flex items-center justify-center h-screen">
-      <p>Coming Soon</p>
-    </div>
-  )
-}
-
-export default HelpPage
